+++ activesupport/test/cache/behaviors/cache_increment_decrement_behavior.rb
@@ -8,7 +8,11 @@ module CacheIncrementDecrementBehavior
     assert_equal 2, @cache.read("foo").to_i
     assert_equal 3, @cache.increment("foo")
     assert_equal 3, @cache.read("foo").to_i
-    assert_nil @cache.increment("bar")
+    if @increment_and_decrement_support_initialization
+      assert_equal 0, @cache.increment("bar", 0)
+    else
+      assert_nil @cache.increment("bar")
+    end
   end
 
   def test_decrement
@@ -18,6 +22,11 @@ module CacheIncrementDecrementBehavior
     assert_equal 2, @cache.read("foo").to_i
     assert_equal 1, @cache.decrement("foo")
     assert_equal 1, @cache.read("foo").to_i
-    assert_nil @cache.decrement("bar")
+
+    if @increment_and_decrement_support_initialization
+      assert_equal 0, @cache.increment("bar", 0)
+    else
+      assert_nil @cache.increment("bar")
+    end
   end
 end